functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-genai>=0.1.0
mangum>=0.17.0
pypdf>=5.0.0
python-multipart>=0.0.12
